@@ -2,15 +2,16 @@ name: example
 description: Example app for form_builder_map_field package
 homepage: https://github.com/danvick/form_builder_map_field/tree/master/example
 version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_form_builder: ^3.8.1
+  flutter_form_builder: ^7.0.0-beta.0
   form_builder_map_field:
     path: ../
 
