@@ -5,13 +5,13 @@ author: Danvick Miller <danvickmiller@gmail.com>
 homepage: https://github.com/danvick/form_builder_map_field
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_form_builder: ^3.8.1
-  google_maps_flutter: ^0.5.25+1
+  flutter_form_builder: ^7.0.0-beta.0
+  google_maps_flutter: ^2.0.10
 
 dev_dependencies:
   flutter_test:
