@@ -1,63 +1,185 @@
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
-class FormBuilderMapField extends StatefulWidget {
-  final String attribute;
-  final List<FormFieldValidator> validators;
-  final CameraPosition initialValue;
-  final bool readonly;
-  final InputDecoration decoration;
-  final ValueChanged<CameraPosition> onChanged;
-  final ValueTransformer valueTransformer;
+import 'map_field_dialog.dart';
 
-  // final GoogleMapController controller;
+class FormBuilderMapField extends FormBuilderField<CameraPosition> {
+  /// Callback method for when the map is ready to be used.
+  ///
+  /// Used to receive a [GoogleMapController] for this [GoogleMap].
+  final MapCreatedCallback? onMapCreated;
+
+  /// The initial position of the map's camera.
+  final CameraPosition? initialCameraPosition;
+
+  /// True if the map should show a compass when rotated.
+  final bool compassEnabled;
+
+  /// True if the map should show a toolbar when you interact with the map. Android only.
+  final bool mapToolbarEnabled;
+
+  /// Geographical bounding box for the camera target.
+  final CameraTargetBounds cameraTargetBounds;
+
+  /// Type of map tiles to be rendered.
+  final MapType mapType;
+
+  /// Preferred bounds for the camera zoom level.
+  ///
+  /// Actual bounds depend on map data and device.
+  final MinMaxZoomPreference minMaxZoomPreference;
+
+  /// True if the map view should respond to rotate gestures.
+  final bool rotateGesturesEnabled;
+
+  /// True if the map view should respond to scroll gestures.
+  final bool scrollGesturesEnabled;
+
+  /// True if the map view should show zoom controls. This includes two buttons
+  /// to zoom in and zoom out. The default value is to show zoom controls.
+  ///
+  /// This is only supported on Android. And this field is silently ignored on iOS.
+  final bool zoomControlsEnabled;
+
+  /// True if the map view should respond to zoom gestures.
+  final bool zoomGesturesEnabled;
+
+  /// True if the map view should be in lite mode. Android only.
+  ///
+  /// See https://developers.google.com/maps/documentation/android-sdk/lite#overview_of_lite_mode for more details.
+  final bool liteModeEnabled;
+
+  /// True if the map view should respond to tilt gestures.
+  final bool tiltGesturesEnabled;
+
+  /// Padding to be set on map. See https://developers.google.com/maps/documentation/android-sdk/map#map_padding for more details.
+  final EdgeInsets padding;
+
+  /// Markers to be placed on the map.
+  final Set<Marker> markers;
+
+  /// Polygons to be placed on the map.
+  final Set<Polygon> polygons;
+
+  /// Polylines to be placed on the map.
+  final Set<Polyline> polylines;
+
+  /// Circles to be placed on the map.
+  final Set<Circle> circles;
+
+  /// Called when the camera starts moving.
+  ///
+  /// This can be initiated by the following:
+  /// 1. Non-gesture animation initiated in response to user actions.
+  ///    For example: zoom buttons, my location button, or marker clicks.
+  /// 2. Programmatically initiated animation.
+  /// 3. Camera motion initiated in response to user gestures on the map.
+  ///    For example: pan, tilt, pinch to zoom, or rotate.
+  final VoidCallback? onCameraMoveStarted;
+
+  /// Called repeatedly as the camera continues to move after an
+  /// onCameraMoveStarted call.
+  ///
+  /// This may be called as often as once every frame and should
+  /// not perform expensive operations.
+  final CameraPositionCallback? onCameraMove;
+
+  /// Called when camera movement has ended, there are no pending
+  /// animations and the user has stopped interacting with the map.
+  final VoidCallback? onCameraIdle;
+
+  /// Called every time a [GoogleMap] is tapped.
+  final ArgumentCallback<LatLng>? onTap;
+
+  /// Called every time a [GoogleMap] is long pressed.
+  final ArgumentCallback<LatLng>? onLongPress;
+
+  /// True if a "My Location" layer should be shown on the map.
+  ///
+  /// This layer includes a location indicator at the current device location,
+  /// as well as a My Location button.
+  /// * The indicator is a small blue dot if the device is stationary, or a
+  /// chevron if the device is moving.
+  /// * The My Location button animates to focus on the user's current location
+  /// if the user's location is currently known.
+  ///
+  /// Enabling this feature requires adding location permissions to both native
+  /// platforms of your app.
+  /// * On Android add either
+  /// `<uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />`
+  /// or `<uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />`
+  /// to your `AndroidManifest.xml` file. `ACCESS_COARSE_LOCATION` returns a
+  /// location with an accuracy approximately equivalent to a city block, while
+  /// `ACCESS_FINE_LOCATION` returns as precise a location as possible, although
+  /// it consumes more battery power. You will also need to request these
+  /// permissions during run-time. If they are not granted, the My Location
+  /// feature will fail silently.
+  /// * On iOS add a `NSLocationWhenInUseUsageDescription` key to your
+  /// `Info.plist` file. This will automatically prompt the user for permissions
+  /// when the map tries to turn on the My Location layer.
+  final bool myLocationEnabled;
+
+  /// Enables or disables the my-location button.
+  ///
+  /// The my-location button causes the camera to move such that the user's
+  /// location is in the center of the map. If the button is enabled, it is
+  /// only shown when the my-location layer is enabled.
+  ///
+  /// By default, the my-location button is enabled (and hence shown when the
+  /// my-location layer is enabled).
+  ///
+  /// See also:
+  ///   * [myLocationEnabled] parameter.
+  final bool myLocationButtonEnabled;
+
+  /// Enables or disables the indoor view from the map
+  final bool indoorViewEnabled;
+
+  /// Enables or disables the traffic layer of the map
+  final bool trafficEnabled;
+
+  /// Enables or disables showing 3D buildings where available
+  final bool buildingsEnabled;
+
+  /// Which gestures should be consumed by the map.
+  ///
+  /// It is possible for other gesture recognizers to be competing with the map on pointer
+  /// events, e.g if the map is inside a [ListView] the [ListView] will want to handle
+  /// vertical drags. The map will claim gestures that are recognized by any of the
+  /// recognizers on this list.
+  ///
+  /// When this set is empty or null, the map will only handle pointer events for gestures that
+  /// were not claimed by any other gesture recognizer.
+  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+
+  //TODO: Add documentation
   final IconData markerIcon;
   final double markerIconSize;
   final Color markerIconColor;
-  final MapType mapType;
   final double height;
-  final bool myLocationButtonEnabled; // widget.myLocationButtonEnabled,
-  final bool myLocationEnabled; // widget.myLocationEnabled,
-  final bool zoomGesturesEnabled; // widget.zoomGesturesEnabled,
-  final Set<Marker> markers; // widget.markers,
-  final void Function(LatLng) onTap; // widget.onTap,
-  final EdgeInsets padding; // widget.padding,
-  final bool buildingsEnabled; // widget.buildingsEnabled,
-  final CameraTargetBounds cameraTargetBounds; // widget.cameraTargetBounds,
-  final Set<Circle> circles; // widget.circles,
-  final bool compassEnabled; // widget.compassEnabled,
-  final Set<Factory<OneSequenceGestureRecognizer>>
-      gestureRecognizers; // widget.gestureRecognizers,
-  final bool indoorViewEnabled; // widget.indoorViewEnabled,
-  final bool mapToolbarEnabled; // widget.mapToolbarEnabled,
-  final MinMaxZoomPreference
-      minMaxZoomPreference; //widget.minMaxZoomPreference,
-  final void Function() onCameraIdle; // widget.onCameraIdle,
-  final void Function() onCameraMoveStarted; // widget.onCameraMoveStarted,
-  final void Function(LatLng) onLongPress; // widget.onLongPress,
-  final Set<Polygon> polygons; // widget.polygons,
-  final Set<Polyline> polylines; // widget.polylines,
-  final bool rotateGesturesEnabled; // widget.rotateGesturesEnabled,
-  final bool scrollGesturesEnabled; // widget.scrollGesturesEnabled,
-  final bool tiltGesturesEnabled; // widget.tiltGesturesEnabled,
-  final bool trafficEnabled; // widget.trafficEnabled,
+  final bool allowClear;
+  final Widget resetIcon;
+  final TextEditingController? controller;
 
-  const FormBuilderMapField({
-    Key key,
-    @required this.attribute,
-    this.validators,
-    this.initialValue,
-    this.readonly = false,
-    this.decoration,
-    this.onChanged,
-    this.valueTransformer,
-    // this.controller,
-    this.markerIcon = Icons.person_pin_circle,
+  FormBuilderMapField({
+    Key? key,
+    //From Super
+    required String name,
+    FormFieldValidator<CameraPosition>? validator,
+    InputDecoration decoration = const InputDecoration(),
+    ValueChanged<CameraPosition?>? onChanged,
+    ValueTransformer<CameraPosition?>? valueTransformer,
+    bool enabled = true,
+    FormFieldSetter<CameraPosition>? onSaved,
+    AutovalidateMode autovalidateMode = AutovalidateMode.disabled,
+    VoidCallback? onReset,
+    FocusNode? focusNode,
+    this.allowClear = true,
+    this.resetIcon = const Icon(Icons.close),
+    this.markerIcon = Icons.person_pin_circle_sharp,
     this.markerIconSize = 30,
     this.markerIconColor = Colors.black,
     this.height = 300,
@@ -72,141 +194,159 @@ class FormBuilderMapField extends StatefulWidget {
     this.tiltGesturesEnabled = true,
     this.myLocationEnabled = false,
     this.myLocationButtonEnabled = true,
-    this.padding = const EdgeInsets.all(0),
+    this.padding = EdgeInsets.zero,
     this.indoorViewEnabled = false,
     this.trafficEnabled = false,
     this.buildingsEnabled = true,
-    this.markers,
+    this.zoomControlsEnabled = true,
+    this.liteModeEnabled = false,
+    this.markers = const {},
     this.onTap,
-    this.circles,
-    this.gestureRecognizers,
+    this.circles = const {},
+    this.gestureRecognizers = const {},
     this.onCameraIdle,
     this.onCameraMoveStarted,
     this.onLongPress,
-    this.polygons,
-    this.polylines,
-  }) : super(key: key);
+    this.polygons = const {},
+    this.polylines = const {},
+    this.onMapCreated,
+    this.initialCameraPosition,
+    this.onCameraMove,
+    this.controller,
+  }) : super(
+          key: key,
+          initialValue: initialCameraPosition,
+          name: name,
+          validator: validator,
+          valueTransformer: valueTransformer,
+          onChanged: onChanged,
+          autovalidateMode: autovalidateMode,
+          onSaved: onSaved,
+          enabled: enabled,
+          onReset: onReset,
+          decoration: decoration,
+          builder: (FormFieldState<CameraPosition> field) {
+            final state = field as _FormBuilderMapFieldState;
+            final InputDecoration effectiveDecoration = decoration
+                .applyDefaults(Theme.of(field.context).inputDecorationTheme);
+
+            return TextField(
+              decoration: effectiveDecoration.copyWith(
+                errorText: field.errorText,
+                suffixIcon: state.shouldShowClearIcon(effectiveDecoration)
+                    ? IconButton(
+                        icon: resetIcon,
+                        onPressed: state.clear,
+                      )
+                    : null,
+              ),
+              enabled: enabled,
+              // Setting readOnly to be true hides the keyboard
+              readOnly: true,
+              controller: state._textFieldController,
+              focusNode: state.effectiveFocusNode,
+              // style: style,
+              // autofocus: autofocus,
+            );
+          },
+        );
 
   @override
   _FormBuilderMapFieldState createState() => _FormBuilderMapFieldState();
 }
 
-class _FormBuilderMapFieldState extends State<FormBuilderMapField> {
-  bool _readonly = false;
-  final GlobalKey<FormFieldState> _fieldKey = GlobalKey<FormFieldState>();
-  final GlobalKey _markerKey = GlobalKey();
-  Completer<GoogleMapController> _controllerCompleter = Completer();
-  FormBuilderState _formState;
+class _FormBuilderMapFieldState
+    extends FormBuilderFieldState<FormBuilderMapField, CameraPosition> {
+  late TextEditingController _textFieldController;
+
+  String get valueString => value?.target.toString() ?? '';
 
   @override
   void initState() {
-    _formState = FormBuilder.of(context);
-    _formState?.registerFieldKey(widget.attribute, _fieldKey);
-    _readonly = (_formState?.readOnly == true) ? true : widget.readonly;
     super.initState();
+    _textFieldController = widget.controller ?? TextEditingController();
+    effectiveFocusNode.addListener(_handleFocus);
   }
 
   @override
-  Widget build(BuildContext context) {
-    // print(_markerKey.currentContext.size);
-    return FormField(
-      key: _fieldKey,
-      enabled: !_readonly,
-      initialValue: widget.initialValue,
-      validator: (val) {
-        for (int i = 0; i < widget.validators.length; i++) {
-          if (widget.validators[i](val) != null)
-            return widget.validators[i](val);
-        }
-        return null;
-      },
-      onSaved: (val) {
-        if (widget.valueTransformer != null) {
-          var transformed = widget.valueTransformer(val);
-          FormBuilder.of(context)
-              ?.setAttributeValue(widget.attribute, transformed);
-        } else
-          _formState?.setAttributeValue(widget.attribute, val);
-      },
-      builder: (FormFieldState<CameraPosition> field) {
-        return InputDecorator(
-          decoration: widget.decoration.copyWith(
-            enabled: !_readonly,
-            errorText: field.errorText,
-          ),
-          child: Container(
-            height: widget.height,
-            child: LayoutBuilder(
-              builder: (context, constraints) {
-                var maxWidth = constraints.biggest.width;
-                var maxHeight = constraints.biggest.height;
-
-                return Stack(
-                  children: <Widget>[
-                    Container(
-                      height: maxHeight,
-                      width: maxWidth,
-                      child: GoogleMap(
-                        initialCameraPosition: widget.initialValue ??
-                            CameraPosition(
-                              target:
-                                  LatLng(37.42796133580664, -122.085749655962),
-                              zoom: 14.4746,
-                            ),
-                        onMapCreated: (GoogleMapController controller) {
-                          if (!_controllerCompleter.isCompleted)
-                            _controllerCompleter.complete(controller);
-                        },
-                        onCameraMove: (CameraPosition newPosition) {
-                          field.didChange(newPosition);
-                          if (widget.onChanged != null)
-                            widget.onChanged(newPosition);
-                        },
-                        mapType: widget.mapType,
-                        myLocationButtonEnabled: widget.myLocationButtonEnabled,
-                        myLocationEnabled: widget.myLocationEnabled,
-                        zoomGesturesEnabled: widget.zoomGesturesEnabled,
-                        markers: widget.markers,
-                        onTap: widget.onTap,
-                        padding: widget.padding,
-                        buildingsEnabled: widget.buildingsEnabled,
-                        cameraTargetBounds: widget.cameraTargetBounds,
-                        circles: widget.circles,
-                        compassEnabled: widget.compassEnabled,
-                        gestureRecognizers: widget.gestureRecognizers,
-                        indoorViewEnabled: widget.indoorViewEnabled,
-                        mapToolbarEnabled: widget.mapToolbarEnabled,
-                        minMaxZoomPreference: widget.minMaxZoomPreference,
-                        onCameraIdle: widget.onCameraIdle,
-                        onCameraMoveStarted: widget.onCameraMoveStarted,
-                        onLongPress: widget.onLongPress,
-                        polygons: widget.polygons,
-                        polylines: widget.polylines,
-                        rotateGesturesEnabled: widget.rotateGesturesEnabled,
-                        scrollGesturesEnabled: widget.scrollGesturesEnabled,
-                        tiltGesturesEnabled: widget.tiltGesturesEnabled,
-                        trafficEnabled: widget.trafficEnabled,
-                      ),
-                    ),
-                    Positioned(
-                      bottom: maxHeight / 2,
-                      right: (maxWidth - widget.markerIconSize) / 2,
-                      child: Container(
-                        key: _markerKey,
-                        child: Icon(
-                          widget.markerIcon,
-                          size: widget.markerIconSize,
-                          color: widget.markerIconColor,
-                        ),
-                      ),
-                    ),
-                  ],
-                );
-              },
-            ),
-          ),
-        );
-      },
-    );
+  void dispose() {
+    // Dispose the _textFieldController when initState created it
+    if (null == widget.controller) {
+      _textFieldController.dispose();
+    }
+    super.dispose();
   }
+
+  Future<void> _handleFocus() async {
+    if (effectiveFocusNode.hasFocus && widget.enabled) {
+      await Future.microtask(
+          () => FocusScope.of(context).requestFocus(FocusNode()));
+      final newValue = await showDialog<CameraPosition>(
+        context: context,
+        builder: (context) => LocationFieldDialog(
+          initialCameraPosition: value ?? widget.initialCameraPosition,
+          onTap: widget.onTap,
+          buildingsEnabled: widget.buildingsEnabled,
+          padding: widget.padding,
+          cameraTargetBounds: widget.cameraTargetBounds,
+          circles: widget.circles,
+          compassEnabled: widget.compassEnabled,
+          gestureRecognizers: widget.gestureRecognizers,
+          indoorViewEnabled: widget.indoorViewEnabled,
+          mapToolbarEnabled: widget.mapToolbarEnabled,
+          mapType: widget.mapType,
+          markerIcon: widget.markerIcon,
+          markerIconColor: widget.markerIconColor,
+          markerIconSize: widget.markerIconSize,
+          markers: widget.markers,
+          minMaxZoomPreference: widget.minMaxZoomPreference,
+          myLocationButtonEnabled: widget.myLocationButtonEnabled,
+          myLocationEnabled: widget.myLocationEnabled,
+          onCameraIdle: widget.onCameraIdle,
+          onCameraMoveStarted: widget.onCameraMoveStarted,
+          onLongPress: widget.onLongPress,
+          polygons: widget.polygons,
+          polylines: widget.polylines,
+          rotateGesturesEnabled: widget.rotateGesturesEnabled,
+          scrollGesturesEnabled: widget.scrollGesturesEnabled,
+          tiltGesturesEnabled: widget.tiltGesturesEnabled,
+          trafficEnabled: widget.trafficEnabled,
+          zoomGesturesEnabled: widget.zoomGesturesEnabled,
+          liteModeEnabled: widget.liteModeEnabled,
+          zoomControlsEnabled: widget.zoomControlsEnabled,
+          onCameraMove: widget.onCameraMove,
+          onMapCreated: widget.onMapCreated,
+        ),
+      );
+      if (newValue != null) {
+        didChange(newValue);
+      }
+    }
+  }
+
+  @override
+  void didChange(CameraPosition? value) {
+    super.didChange(value);
+    _textFieldController.text =
+        widget.valueTransformer?.call(value)?.toString() ?? valueString;
+    widget.onChanged?.call(value);
+  }
+
+  void clear() async {
+    _hideKeyboard();
+    // Fix for ripple effect throwing exception
+    // and the field staying gray.
+    // https://github.com/flutter/flutter/issues/36324
+    WidgetsBinding.instance?.addPostFrameCallback((_) {
+      setState(() => _textFieldController.clear());
+    });
+  }
+
+  void _hideKeyboard() {
+    Future.microtask(() => FocusScope.of(context).requestFocus(FocusNode()));
+  }
+
+  bool shouldShowClearIcon([InputDecoration? decoration]) =>
+      (_textFieldController.text.isNotEmpty || effectiveFocusNode.hasFocus) &&
+      decoration?.suffixIcon == null;
 }
