@@ -1,4 +1,5 @@
 library form_builder_map_field;
 
-export './src/form_builder_map_field.dart';
+export 'src/form_builder_map_field.dart';
+export 'src/map_field_dialog.dart';
 export 'package:google_maps_flutter/google_maps_flutter.dart';
